+++ app/Http/Controllers/Admin/Categories/SubcategoriesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin\Categories;
 
 use App\Category;
+use App\Http\Controllers\Helpers\DirectoryEditor;
 use App\Http\Controllers\Helpers\Helpers;
 use App\Http\Controllers\Helpers\Validation;
 use App\Subcategory;
@@ -86,13 +87,17 @@ class SubcategoriesController extends MainCategoriesController
             if ($request->subcategoryId) {
                 $ids[] = $request->subcategoryId;
             }
-            if (Subcategory::whereIn('id', $ids)->delete()) {
-                return response(
-                    [
-                        'error' => false,
-                        'ids' => $ids
-                    ]
-                );
+
+            $deleteDir = DirectoryEditor::clearAfterSubcategoryDelete($ids);
+            if (!$deleteDir['error']) {
+                if (Subcategory::whereIn('id', $ids)->delete()) {
+                    return response(
+                        [
+                            'error' => false,
+                            'ids' => $ids
+                        ]
+                    );
+                };
             }
         } catch (\Exception $e) {
             return response(
